% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savedataR.R
\name{savedataR}
\alias{savedataR}
\title{Save Data and Log Summary Statistics}
\usage{
savedataR(data, csv_file, log_file)
}
\arguments{
\item{data}{A data frame to be saved.}

\item{csv_file}{Path to the output CSV file.}

\item{log_file}{Path to the log file where summary statistics will be written.}
}
\value{
None.
}
\description{
This function saves a data frame to a CSV file and logs summary statistics
to a specified log file.
}
\examples{
savedataR(mtcars, "output.csv", "log.txt")

}
